I had to change the number of mines in each game. The number of mines initialized in each board constructor must be smaller than the # of rows and columns.
SimpleBoard was fine because it had 5 mines and 7 rows/columns.

MediumBoard was wrong because it had 10 mines and 10 rows/columns, so it was running out of numbers to choose from in shuffle().

Same for AdvBoard. It had 17 mines, but only 13 rows/columns. It was running out of numbers to choose from in shuffle().

I changed the number of mines in both AdvBoard and MediumBoard to a smaller number.

Signed-off-by: Laura Flores <laura-flores@uiowa.edu>